samples/multimodal/retrievers/firestore/requirements.txt
google-cloud==0.34.0
pillow==10.4.0
langchain-google-firestore==0.3.0
langchain-google-vertexai==1.0.10